popravljeno atof za gain preko komandne linije za sva tri modela

+++ Day_1_model2/aadsp2_projekat/model_1/main.cpp
@@ -149,7 +149,7 @@ int main(int argc, char* argv[])
 	strcpy(WavOutputName, argv[2]);
 	wav_out = OpenWavFileForRead(WavOutputName, "wb");
 	MODE = atoi(argv[3]);
-	INPUT_GAIN = atoi(argv[4]);
+	INPUT_GAIN = atof(argv[4]);
 	//-------------------------------------------------
 
 	// Read input wav header
